requirements.txt
discord.py
ultralytics==8.3.65
torch==2.5.1
Pillow
python-dotenv